Factor TestRequest out into the test support module.
